+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this mod will be documented in this file.
   - Fixed various bugs in original implementation
   - Added +10% supply use to S-mod stats
 - Hullmod tooltip reworked
+- Added status indicator to ability tooltip's title
 
 ## [0.1.18] - 2026-05-03